[libcommhistory] Do not match contact groups with localUid. Contributes to JB#33949
RecipientList::matches calls Recipient::matches for each recipient. The
Recipient::matches checks localUid only for IM recipients, not for phone numbers.

With multi sim each sim has unique localUid and it should not make
a different which sim card received the message. Whether we have
an existing group for a remoteUid or not is the only thing that
matters. Long story short, the localUid comparison doesn't make sense when
finding matching group.

We might need to change the Recipient::matches in future.