[libcommhistory] Remove trailing slash at the end of RING_ACCOUNT. Contributes to JB#34332
When the messaging application is starting and it is known that
we are using telepathy ring account, it might take a bit of time
that current active modem is resolved.

Further, the active modem is not really needed to be known when
resolving contacts or contact groups. So let's remove the
trailing slash.

Now the RING_ACCOUNT is also in sync with commhistoryd.