[libcommhistory] Make DatabaseIO::getEventByMmsId useful. Contributes to JB#18573
The plan is to use the mmsId field to store the location url
from the MMS push notification, to detect the duplicates.

The DatabaseIO::getEventByMmsId method is apparently not being
used (and it's hard to imagine how it could possibly be used in
Sailfish OS) so it should be ok to change its signature.