[libcommhistory] Use libphonenumber for matching. Contributes to JB#38835
Examples of solved false positive matches:
"+36 20 123 4567" and "+36 30 123 4567" were matched incorrectly.
"+36 20 123 4567" and "+44 20 123 4567" were matched incorrectly.

Examples of local and IDD numbers that still work:
"+36 20 123 4567" still matches "06 20 123 4567" (local)
"+36 20 123 4567" still matches "00 36 70 381 0581" (IID)

Examples of remaining false positives:
"06 20 123 4567" is a false positive match of "+1 0620 123 4567"
"00 36 20 123 4567" is a false positive match of "+1 00 36 20 123 4567"

We deem that the remaining false positives are unlikely enough that
they don't matter in practice. If this becomes relevant, we will need
to make it possible to pass a region code to the matching algorithm.