[libcommhistory] Use standalone test database when running tests. JB#52802
Use a separate test database instead of polluting the general database
when running the tests. This also means the test database can be
freely deleted before running each tests, to ensure stray database
entries from one tests (e.g. left behind if a test crashes) do not
cause failures in following tests when the tests are run in a batch.