[libcommhistory] Use minimized phone number to identify RecipientPrivate instances. Fixes JB#52774
Minimized phone numbers are needed to compare numbers with different
country calling codes correctly. Also in case of phone numbers it should
in practice ignore localUid so instead of keeping it as is use only
RING_ACCOUNT (which is the prefix of localUid for phone numbers) to
identify those RecipientPrivate instances.

This fixes problems with commhistory-daemon unit tests.

It would be better to fix this so that libphonenumber was used to
compare these values and localUid was correctly handled but that would
require greater changes.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>