[libcommhistory] Fix race when sending a temporary file with MMS. JB#56314
If the file to be sent is a temporary file, the owner might assume
it can be removed after sendMessage() finishes. Avoid that by
creating a temporary file owned by the helper, hard link as first
option and file copy if that fails.

The copy case might not be the most efficient, but the files should
be small enough here and it's simple on api.