[libcommhistory] Fix unstable unit test. JB#57536
The deleteTestContact() given a change listener creates an event loop
and continues processing until the change happens.

On the other side SeasideCache was hitting idle processing
to remove instances for contacts being in m_expiredContacts,
and when reaching SeasideCache::contactsRemoved() them being already
gone, and thus change listener not getting a callback with the instance.
And that, on its part, confusing libcommhistory side.

It's up for consideration how much SeasideCache should cope with
extra event loops, but those are rare, likely a bad idea, and regardless
of it, in the unit test there shouldn't be reason for waiting individual
contact deletions to finish before proceeding to the next one.
The QTRY_COMPARE() should ensure that it ends up with expected state.